Feat: implemented a simple .csv parser
Implemented the initial version of the .csv parser. So far, we have a
method for reading the file and saving it in a vector in-memory and for
printing this vector. Future work probably involves using an
unordered_map to save the data in-memory instead of vector and new
methods as needed.